Update ItemController actions

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,29 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
 
-        public ItemController(ApplicationDbContext context)
+        private readonly UserManager<ApplicationUser> _userManager;
+
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            ApplicationUser user = _userManager.GetUserAsync(HttpContext.User).Result;
+            var model = _context.Items.Where(x=>x.User.Id == user.Id).ToList();
 
             return View("Index", model);
         }
@@ -29,6 +37,8 @@ namespace WishList.Controllers
         [HttpPost]
         public IActionResult Create(Models.Item item)
         {
+            ApplicationUser user = _userManager.GetUserAsync(HttpContext.User).Result;
+            item.User = user;
             _context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
@@ -37,6 +47,9 @@ namespace WishList.Controllers
         public IActionResult Delete(int id)
         {
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            ApplicationUser user = _userManager.GetUserAsync(HttpContext.User).Result;
+            if (item.User != user)
+                return Unauthorized();
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
